TAD Lista. Comentarios no TAD Lista e arquivos do TAD Fila.

+++ tad/lista/src/list.c
@@ -145,7 +145,7 @@ void TLista_Remover(TLista* Lista, TListaNo* No)
 	Lista->Tamanho--;
 }
 
-unsigned int TLista_Tamanho(TLista* Lista)
+size_t TLista_Tamanho(TLista* Lista)
 {
 	return Lista->Tamanho;
 }
